+++ dongtai-core/src/main/java/io/dongtai/iast/core/utils/threadlocal/TaintRangesPool.java
@@ -4,17 +4,17 @@ import io.dongtai.iast.core.handler.hookpoint.models.taint.range.TaintRanges;
 
 import java.util.Map;
 
-public class TaintRangesPool extends ThreadLocal<Map<Integer, TaintRanges>> {
+public class TaintRangesPool extends ThreadLocal<Map<Long, TaintRanges>> {
     @Override
-    protected Map<Integer, TaintRanges> initialValue() {
+    protected Map<Long, TaintRanges> initialValue() {
         return null;
     }
 
-    public void add(Integer hash, TaintRanges taintRanges) {
+    public void add(Long hash, TaintRanges taintRanges) {
         this.get().put(hash, taintRanges);
     }
 
-    public TaintRanges get(int hash) {
+    public TaintRanges get(long hash) {
         return this.get().get(hash);
     }
 }
